refactor OnboardingScreen
